@@ -41,57 +41,60 @@
     "compile": "webpack --config=build/webpack.app.config.js --env=production",
     "build": "webpack --config=build/webpack.app.config.js --env=production && electron-builder --publish never",
     "release": "webpack --config=build/webpack.app.config.js --env=production && electron-builder --publish never",
-    "dist": "yarn compile && electron-builder --publish never"
+    "dist": "yarn compile && electron-builder --publish never",
+    "lint": "yarn run eslint --ext .js,.vue src",
+    "lintFix": "yarn run eslint --ext .js,.vue src --fix"
   },
   "dependencies": {
-    "@panter/vue-i18next": "^0.15.1",
-    "bitcoinjs-lib": "^5.1.7",
-    "bitsharesjs": "^3.1.1",
-    "bootstrap-vue": "^2.0.3",
-    "chalk": "^2.4.2",
-    "crypto-js": "^3.1.9-1",
-    "dexie": "^2.0.4",
-    "electron-devtools-installer": "^2.2.4",
-    "elliptic": "^6.5.4",
-    "fs-jetpack": "^2.2.2",
-    "i18next": "^17.1.0",
-    "i18next-node-fs-backend": "^2.1.3",
+    "@noble/ed25519": "^1.6.1",
+    "@noble/secp256k1": "^1.6.3",
+    "balm-ui": "^10.9.3",
+    "bitcoinjs-lib": "^6.0.1",
+    "bitsharesjs": "github:bitshares/bitsharesjs#develop",
+    "crypto-js": "^4.1.1",
+    "dexie": "^3.2.2",
+    "electron-devtools-installer": "^3.2.0",
+    "linebyline": "^1.3.0",
     "lzma": "^2.3.2",
-    "otpauth": "^5.0.3",
-    "readline": "^1.3.0",
-    "typeface-rajdhani": "0.0.72",
-    "typeface-roboto": "0.0.75",
-    "vue": "^2.6.10",
-    "vue-multiselect": "^2.1.6",
-    "vue-password-strength-meter": "^1.5.0",
-    "vue-router": "^3.1.3",
-    "vuex": "^3.1.1",
-    "ws": "^7.5.6",
-    "zxcvbn": "^4.4.2"
+    "mitt": "^3.0.0",
+    "otpauth": "^8.0.1",
+    "query-string": "^7.1.1",
+    "simple-websocket": "^9.1.0",
+    "socket.io": "^4.5.1",
+    "typeface-rajdhani": "1.1.13",
+    "typeface-roboto": "1.1.13",
+    "uuid": "^8.3.2",
+    "vue": "^3.2.37",
+    "vue-i18n": "9",
+    "vue-router": "^4.1.2",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.3",
-    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
-    "@babel/plugin-transform-runtime": "^7.6.2",
-    "@babel/preset-env": "^7.6.3",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.0.6",
-    "css-loader": "^3.2.0",
-    "electron": "^3.0.0",
-    "electron-builder": "^21.2.0",
-    "eslint": "^6.5.1",
-    "eslint-plugin-vue": "^5.2.3",
-    "file-loader": "^4.2.0",
-    "friendly-errors-webpack-plugin": "^1.7.0",
-    "node-sass": "^4.14.1",
-    "sass-loader": "^8.0.0",
-    "source-map-support": "^0.5.13",
-    "style-loader": "^1.0.0",
-    "vue-loader": "^15.7.1",
-    "vue-template-compiler": "^2.6.10",
-    "webpack": "^4.41.0",
-    "webpack-cli": "^3.3.9",
-    "webpack-merge": "^4.2.2",
-    "webpack-node-externals": "^1.7.2"
+    "@babel/core": "^7.18.9",
+    "@babel/plugin-proposal-object-rest-spread": "^7.18.9",
+    "@babel/plugin-transform-runtime": "^7.18.9",
+    "@babel/preset-env": "^7.18.9",
+    "@soda/friendly-errors-webpack-plugin": "^1.8.1",
+    "babel-loader": "^8.2.5",
+    "css-loader": "^6.7.1",
+    "electron": "^19.0.8",
+    "electron-builder": "^23.1.0",
+    "eslint": "^8.20.0",
+    "eslint-plugin-vue": "^9.2.0",
+    "file-loader": "^6.2.0",
+    "sass": "^1.53.0",
+    "sass-loader": "^13.0.2",
+    "source-map-support": "^0.5.21",
+    "style-loader": "^3.3.1",
+    "vue-loader": "^17.0.0",
+    "vue-template-compiler": "^2.7.7",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-merge": "^5.8.0",
+    "webpack-node-externals": "^3.0.0"
+  },
+  "optionalDependencies": {
+    "bufferutil": "^4.0.6",
+    "utf-8-validate": "^5.0.9"
   }
 }
