@@ -1,77 +1,113 @@
-<template>
-    <b-modal
-        id="type"
-        ref="modalComponent"
-        centered
-        no-close-on-esc
-        no-close-on-backdrop
-        hide-header-close
-        hide-footer
-        :title="$t('operations:account_id.title')"
-    >
-        <div
-            v-b-tooltip.hover
-            :title="$t('operations:relink.request_tooltip')"
-        >
-            {{ $t('operations:relink.request', {appName: incoming.appName, origin: incoming.origin, chain: incoming.chain, accountId: beetapp.account_id }) }} &#10068;
-        </div>
-        <br>
-        <b-btn
-            class="mt-3"
-            variant="success"
-            block
-            @click="_clickedAllow"
-        >
-            {{ $t('operations:link.accept_btn') }}
-        </b-btn>
-        <b-btn
-            class="mt-1"
-            variant="danger"
-            block
-            @click="_clickedDeny"
-        >
-            {{ $t('operations:link.reject_btn') }}
-        </b-btn>
-    </b-modal>
-</template>
-<script>
-    import AbstractPopup from "./abstractpopup";
+<script setup>
+    import { ipcRenderer } from 'electron';
+    import { onMounted, computed} from "vue";
+    import { useI18n } from 'vue-i18n';
+    const { t } = useI18n({ useScope: 'global' });
     import RendererLogger from "../../lib/RendererLogger";
     const logger = new RendererLogger();
 
-    export default {
-        name: "ReLinkRequestPopup",
-        components: {},
-        extends: AbstractPopup,
-        data() {
-            return {
-                type: "ReLinkRequestPopup",
-                chosenAccount: { trackId: 0 },
-                beetapp: {}
-            };
+    const props = defineProps({
+        request: {
+            type: Object,
+            required: true,
+            default() {
+                return {}
+            }
         },
-        mounted() {
-            logger.debug("Relink Popup initialised");
-        },
-        methods: {
-            _onShow: function() {
-                this.error = false;
-                console.log("Popup incoming, payload:", this.incoming);
-                this.beetapp = this.$store.state.OriginStore.apps.filter(
-                    x => x.identityhash == this.incoming.payload.identityhash
-                )[0];
-            },
-            _execute: function() {
-                let account = this.$store.state.AccountStore.accountlist.filter(
-                    x => x.accountID == this.beetapp.account_id && x.chain == this.beetapp.chain
-                );
-                return {
-                    identityhash: this.incoming.payload.identityhash,
-                    name: account.accountName,
-                    chain: this.beetapp.chain,
-                    id: this.beetapp.account_id
-                };
+        accounts: {
+            type: Array,
+            required: true,
+            default() {
+                return []
             }
         }
-    };
+    });
+
+    let requestText = computed(() => {
+        if (!props.request || !props.accounts.length) {
+            return '';
+        }
+
+        return t(
+            'operations.relink.request',
+            {
+                appName: props.request.appName,
+                origin: props.request.origin,
+                chain: props.request.chain,
+                accountId: props.accounts[0].accountID
+            }
+        );
+    });
+
+    onMounted(() => {
+        logger.debug("Relink Popup initialised");
+    });
+
+    function _clickedAllow() {
+        ipcRenderer.send(
+            "clickedAllow",
+            {
+                result: {
+                    identityhash: props.request.payload.identityhash,
+                    name: props.accounts[0].accountName,
+                    chain: props.accounts[0].chain,
+                    id: props.accounts[0].accountID,
+                    success: true
+                },
+                request: {
+                    id: props.request.id
+                }
+            }
+        );
+    }
+
+    function _clickedDeny() {
+        ipcRenderer.send(
+            "clickedDeny",
+            {
+                result: {canceled: true},
+                request: {id: props.request.id}
+            }
+        );
+    }
 </script>
+
+<template>
+    <div
+        v-if="props.request && props.accounts"
+        style="padding:5px"
+    >
+        <div v-tooltip="t('operations.relink.request_tooltip')">
+            {{ requestText }} &#10068;
+        </div>
+        <br>
+        <ui-button
+            raised
+            style="margin-right:5px"
+            @click="_clickedAllow()"
+        >
+            {{ t('operations.link.accept_btn') }}
+        </ui-button>
+        <ui-button
+            raised
+            @click="_clickedDeny()"
+        >
+            {{ t('operations.link.reject_btn') }}
+        </ui-button>
+    </div>
+    <div
+        v-else
+        style="padding:5px"
+    >
+        <div>
+            Error showing prompt
+        </div>
+        <br>
+        <ui-button
+            raised
+            @click="_clickedDeny()"
+        >
+            {{ t('operations.link.reject_btn') }}
+        </ui-button>
+    </div>
+</template>
