@@ -1,132 +1,157 @@
-<template>
-    <b-modal
-        id="type"
-        ref="modalComponent"
-        centered
-        no-close-on-esc
-        no-close-on-backdrop
-        hide-header-close
-        hide-footer
-        :title="$t('operations:transfer.title')"
-    >
-        {{ message }}:
-        <br>
-        <br>
-        <pre class="text-left custom-content"><code>
-  Recipient: {{ to }}
-  Amount: {{ toSend }}
-{{ toSendFee != null ? "    Fee: " + toSendFee : "" }}
-        </code></pre>
-        <b-btn
-            class="mt-3"
-            variant="success"
-            block
-            @click="_clickedAllow"
-        >
-            {{ $t("operations:transfer.accept_btn") }}
-        </b-btn>
-        <b-btn
-            class="mt-1"
-            variant="danger"
-            block
-            @click="_clickedDeny"
-        >
-            {{ $t("operations:transfer.reject_btn") }}
-        </b-btn>
-    </b-modal>
-</template>
-<script>
-    import AbstractPopup from "./abstractpopup";
+<script setup>
+    import { ipcRenderer } from 'electron';
+    import { onMounted, computed } from "vue";
     import RendererLogger from "../../lib/RendererLogger";
-    import getBlockchain from "../../lib/blockchains/blockchainFactory";
+    import { useI18n } from 'vue-i18n';
+    const { t } = useI18n({ useScope: 'global' });
     const logger = new RendererLogger();
-    import {getKey} from '../../lib/SecureRemote';
 
-    export default {
-        name: "TransferRequestPopup",
-        extends: AbstractPopup,
-        data() {
-            return {
-                type: "TransferRequestPopup",
-                askWhitelist: false,
-                message: null,
-                recipient: null,
-                satoshis: null,
-                feeInSatoshis: null,
-                asset_id: null,
-                toSend: null,
-                toSendFee: null
-            };
+    const props = defineProps({
+        request: {
+            type: Object,
+            required: true,
+            default() {
+                return {}
+            }
         },
-        methods: {
-            _onShow() {
-                this.message = this.$t("operations:transfer.request", {
-                    appName: this.incoming.appName,
-                    origin: this.incoming.origin,
-                    chain:   this.$store.getters['AccountStore/getSigningKey'](this.incoming).chain,
-                    accountName:   this.$store.getters['AccountStore/getSigningKey'](this.incoming).accountName
-                });
-
-                let blockchain = getBlockchain(this.incoming.chain);
-
-                this.to = this.incoming.params.to;
-                this.satoshis = this.incoming.params.amount.satoshis;
-                this.asset_id = this.incoming.params.amount.asset_id;
-
-                this.toSend = blockchain.format(
-                    this.incoming.params.amount
-                );
-
-                if (blockchain.supportsFeeCalculation()) {
-                    this.toSendFee = "Calculating ...";
-                    let loadFee = async () => {
-                        try {
-                            let result = await blockchain.transfer(
-                                await getKey(this.$store.getters['AccountStore/getSigningKey'](this.incoming).keys.active),
-                                this.$store.getters['AccountStore/getSigningKey'](this.incoming).accountName,
-                                this.incoming.params.to,
-                                {
-                                    amount: this.incoming.params.amount.satoshis || this.incoming.params.amount.amount,
-                                    asset_id: this.incoming.params.amount.asset_id
-                                },
-                                this.incoming.params.memo,
-                                false
-                            );
-                            this.feeInSatoshis = result.feeInSatoshis;
-                            this.toSendFee = blockchain.format(
-                                this.feeInSatoshis
-                            );
-                        } catch (err) {
-                            console.error(err);
-                            this.toSendFee = err;
-                        }
-                    };
-                    loadFee().then(()=>{}).catch(()=>{});
-                } else {
-                    this.toSendFee = null;
-                }
-            },
-            getSuccessNotification(result) {
-                return {msg: 'Transaction `transfer` successfully broadcast.', link: '' };
-            },
-            _execute: async function () {
-                let blockchain = getBlockchain(this.incoming.chain);
-
-                if (!this.incoming.params.amount) {
-                    this.incoming.params.amount = this.incoming.params.satoshis;
-                }
-
-                return await blockchain.transfer(
-                    await getKey(this.$store.getters['AccountStore/getSigningKey'](this.incoming).keys.active),
-                    this.$store.getters['AccountStore/getSigningKey'](this.incoming).accountName,
-                    this.incoming.params.to,
-                    {
-                        amount: this.incoming.params.amount.satoshis || this.incoming.params.amount.amount,
-                        asset_id: this.incoming.params.amount.asset_id
-                    },
-                    this.incoming.params.memo,
-                );
+        chain: {
+            type: String,
+            required: true,
+            default() {
+                return ''
+            }
+        },
+        accountName: {
+            type: String,
+            required: true,
+            default() {
+                return ''
+            }
+        },
+        toSend: {
+            type: String,
+            required: true,
+            default() {
+                return ''
             }
         }
-    };
+    });
+
+    let message = computed(() => {
+        if (!props.request || !props.chain || !props.accountName) {
+            return '';
+        }
+        return t("operations.transfer.request", {
+            appName: props.request.payload.appName,
+            origin: props.request.payload.origin,
+            chain: props.chain,
+            accountName: props.accountName
+        });
+    });
+
+    let to = computed(() => {
+        if (!props.request) {
+            return '';
+        }
+        return props.request.payload.params.to;
+    });
+
+    let satoshis = computed(() => {
+        if (!props.request) {
+            return '';
+        }
+        return props.request.payload.params.amount.satoshis;
+    });
+
+    let asset_id = computed(() => {
+        if (!props.request) {
+            return '';
+        }
+        return props.request.payload.params.amount.asset_id;
+    });
+
+    let toSend = computed(() => {
+        if (!props.toSend) {
+            return '';
+        }
+        return props.toSend;
+    });
+
+    let toSendFee = computed(() => {
+        if (!props.request) {
+            return '';
+        }
+        return props.request.payload.params.toSendFee ?? null;
+    });
+
+    let feeInSatoshis = computed(() => {
+        if (!props.request) {
+            return '';
+        }
+        return props.request.payload.params.feeInSatoshis ?? null;
+    });
+
+    onMounted(() => {
+        logger.debug("Transfer request popup initialised");
+    });
+
+    function _clickedAllow() {
+        ipcRenderer.send(
+            "clickedAllow",
+            {
+                result: {success: true},
+                request: {id: props.request.id}
+            }
+        );
+    }
+
+    function _clickedDeny() {
+        ipcRenderer.send(
+            "clickedDeny",
+            {
+                result: {canceled: true},
+                request: {id: props.request.id}
+            }
+        );
+    }
 </script>
+<template>
+    <div style="padding:5px">
+        <Text>
+            {{ message }}
+        </Text>
+        <ui-list>
+            <ui-item key="Recipient">
+                <ui-item-text-content>
+                    Recipient: {{ to }}
+                </ui-item-text-content>
+            </ui-item>
+            <ui-item key="Amount">
+                <ui-item-text-content>
+                    Amount: {{ toSend }}
+                </ui-item-text-content>
+            </ui-item>
+            <ui-item
+                v-if="toSendFee"
+                key="Fee"
+            >
+                <ui-item-text-content>
+                    Fee: {{ toSendFee }}
+                </ui-item-text-content>
+            </ui-item>
+        </ui-list>
+        <ui-button
+            raised
+            style="margin-right:5px"
+            @click="_clickedAllow()"
+        >
+            {{ t("operations.transfer.accept_btn") }}
+        </ui-button>
+        <ui-button
+            raised
+            @click="_clickedDeny()"
+        >
+            {{ t("operations.transfer.reject_btn") }}
+        </ui-button>
+    </div>
+</template>
