@@ -2,9 +2,7 @@ import BlockchainAPI from "./BlockchainAPI";
 import RendererLogger from "../RendererLogger";
 const logger = new RendererLogger();
 
-const fetch = require('node-fetch');
-
-const { Api, JsonRpc, RpcError } = require('eosjs');
+import { Api, JsonRpc, RpcError } from 'eosjs';
 import { JsSignatureProvider } from "eosjs/dist/eosjs-jssig";
 
 import * as ecc from "eosjs-ecc";
