@@ -1,37 +1,31 @@
-<template>
-    <div class="top">
-        <div class="container-fluid">
-            <div class="row">
-                <div class="col-12 d-flex smallHeader">
-                    <img
-                        class="logo-small"
-                        src="img/beet.png"
-                        alt
-                    >
-                    <h4 class="h4 beet-typo-small">
-                        {{ $t('appName') }}
-                    </h4>
-                    <LangSelect ref="langswitch" />
-                </div>
-            </div>
-        </div>
-    </div>
-</template>
-
-<script>
-    import LangSelect from "./lang-select";
+<script setup>
+    import { onMounted } from 'vue';
+    import { useI18n } from 'vue-i18n';
+    const { t } = useI18n({ useScope: 'global' });
     import RendererLogger from "../lib/RendererLogger";
     const logger = new RendererLogger();
 
-    export default {
-        name: "HeaderSmall",
-        i18nOptions: { namespaces: "common" },
-        components: { LangSelect },
-        data() {
-            return {};
-        },
-        mounted() {
-            logger.debug("Small Header mounted");
-        }
-    };
+    onMounted(() => {
+        logger.debug("Small Header mounted");
+    });
 </script>
+
+<template>
+    <div class="top">
+        <ui-grid class="container-fluid">
+            <ui-grid-cell
+                class="smallHeader"
+                columns="12"
+            >
+                <img
+                    class="logo-small"
+                    src="img/beet.png"
+                    alt
+                >
+                <h4 class="h4 beet-typo-small">
+                    {{ t('common.appName') }}
+                </h4>
+            </ui-grid-cell>
+        </ui-grid>
+    </div>
+</template>
