@@ -1,92 +1,105 @@
-<template>
-    <b-modal
-        id="type"
-        ref="modalComponent"
-        centered
-        no-close-on-esc
-        no-close-on-backdrop
-        hide-header-close
-        hide-footer
-        :title="incoming.generic.title"
-    >
-        {{ incoming.generic.message }}:
-        <br>
-        <br>
-        <pre class="text-left custom-content"><code>{{ incoming.generic.details }}</code></pre>
-        <b-btn
-            class="mt-3"
-            variant="success"
-            block
-            @click="_clickedAllow"
-        >
-            {{ incoming.generic.acceptText || $t('operations:rawsig.accept_btn') }}
-        </b-btn>
-        <b-btn
-            class="mt-1"
-            variant="danger"
-            block
-            @click="_clickedDeny"
-        >
-            {{ incoming.generic.rejectText || $t('operations:rawsig.reject_btn') }}
-        </b-btn>
-    </b-modal>
-</template>
-<script>
-    import AbstractPopup from "./abstractpopup";
+<script setup>
+    import { ipcRenderer } from 'electron';
+    import { onMounted, computed } from "vue";
+    import { useI18n } from 'vue-i18n';
     import RendererLogger from "../../lib/RendererLogger";
-    import getBlockchain from "../../lib/blockchains/blockchainFactory";
-    import {getKey} from '../../lib/SecureRemote';
+
+    const { t } = useI18n({ useScope: 'global' });
     const logger = new RendererLogger();
 
-    export default {
-        name: "GenericRequestPopup",
-        extends: AbstractPopup,
-        data() {
-            return {
-                type: "GenericRequestPopup",
-                incoming: {
-                    generic: {}
-                }
-            };
-        },
-        mounted() {
-            logger.debug("Req Popup initialised");
-        },
-        methods: {
-            _execute: async function () {
-                let returnValue = null;
-                if (this.incoming.acceptCall) {
-                    returnValue = this.incoming.acceptCall();
-                } else {
-                    let blockchain = getBlockchain(this.incoming.chain);
-                    let operation = await blockchain.getOperation(
-                        this.incoming,
-                        this._getLinkedAccount()
-                    );
-                    if (!operation.notingToDo) {
-                        let transaction = await blockchain.sign(
-                            operation,
-                            await getKey(this.$store.getters['AccountStore/getSigningKey'](this.incoming).keys.active)
-                        );
-                        returnValue = await blockchain.broadcast(transaction);
-                    } else {
-                        returnValue = {
-                            msg: "Already done, no action needed"
-                        }
-                    }
-                }
-                if (this.allowWhitelist) {
-                    // todo: allowWhitelist move whitelisting into BeetAPI
-                    this.$store.dispatch(
-                        "WhitelistStore/addWhitelist",
-                        {
-                            identityhash: this.incoming.identityhash,
-                            method: 'signMessage'
-                        }
-                    );
-                }
-                return returnValue;
+    const props = defineProps({
+        request: {
+            type: Object,
+            required: true,
+            default() {
+                return {}
             }
+        },
+        payload: {
+            type: Object,
+            required: true,
+            default() {
+                return {}
+            }
+        },
+    });
+
+    let acceptText = computed(() => {
+        if (!props.payload) {
+            return '';
         }
-    };
+        return props.payload.generic.acceptText ?? t('operations.rawsig.accept_btn');
+    });
+
+    let rejectText = computed(() => {
+        if (!props.payload) {
+            return '';
+        }
+        return props.payload.generic.rejectText ?? t('operations.rawsig.reject_btn');
+    });
+
+    let votePrompt = computed(() => {
+        if (!props.payload) {
+            return '';
+        }
+        return props.payload.generic.details.split(/\r?\n/);
+    });
+
+    onMounted(() => {
+        logger.debug("Req Popup initialised");
+    });
+
+    function _clickedAllow() {
+        ipcRenderer.send(
+            "clickedAllow",
+            {
+                result: {success: true},
+                request: {id: props.request.id}
+            }
+        );
+    }
+
+    function _clickedDeny() {
+        ipcRenderer.send(
+            "clickedDeny",
+            {
+                result: {canceled: true},
+                request: {id: props.request.id}
+            }
+        );
+    }
 </script>
+
+<template>
+    <div style="padding:5px">
+        <Text size="md">
+            {{ props.payload.generic.message }}:
+        </Text>
+
+        <ui-list>
+            <ui-item
+                v-for="item in votePrompt"
+                :key="item"
+            >
+                <ui-item-text-content>
+                    {{ item }}
+                </ui-item-text-content>
+            </ui-item>
+        </ui-list>
+
+        <ui-button
+            raised
+            style="margin-right:5px"
+            @click="_clickedAllow()"
+        >
+            {{ acceptText }}
+        </ui-button>
+
+        <ui-button
+            outlined
+            @click="_clickedDeny()"
+        >
+            {{ rejectText }}
+        </ui-button>
+    </div>
+</template>
