@@ -1,187 +1,160 @@
+<script setup>
+    import { onMounted, watchEffect, watch, ref, computed } from 'vue';
+    import { useI18n } from 'vue-i18n';
+    import { ipcRenderer } from "electron";
+    import store from '../store/index';
+    import RendererLogger from "../lib/RendererLogger";
+    import {formatAccount} from "../lib/formatter";
+
+    const { t } = useI18n({ useScope: 'global' });
+    const logger = new RendererLogger();
+    let tableData = ref();
+
+    onMounted(() => {
+        logger.debug("Settings Mounted");
+    });
+
+    function fetchDapps() {
+        let storedDapps = [];
+        for (let i = 0; i < store.state.AccountStore.accountlist.length; i++) {
+            let apps = store.getters['OriginStore/walletAccessibleDapps'](store.state.AccountStore.accountlist[i].accountID, store.state.AccountStore.accountlist[i].chain);
+            if (typeof apps != 'undefined') {
+                storedDapps = storedDapps.concat(apps);
+            }
+        }
+        return storedDapps;
+    }
+
+    let dapps = computed(() => {
+        let storedDapps = [];
+        for (let i = 0; i < store.state.AccountStore.accountlist.length; i++) {
+            let apps = store.getters['OriginStore/walletAccessibleDapps'](store.state.AccountStore.accountlist[i].accountID, store.state.AccountStore.accountlist[i].chain);
+            if (typeof apps != 'undefined') {
+                storedDapps = storedDapps.concat(apps);
+            }
+        }
+        return storedDapps;
+    })
+
+    function getDisplayString(accountID, chain) {
+        let account = store.getters['AccountStore/getSafeAccount']({account_id: accountID, chain: chain});
+        return formatAccount(account);
+    }
+
+    watchEffect(() => {
+        if (dapps.value && dapps.value.length) {
+            tableData.value = {
+                data: dapps.value.map(dapp => {
+                    return {
+                        appName: dapp.appName,
+                        origin: dapp.origin,
+                        displayString: getDisplayString(dapp.account_id, dapp.chain),
+                        chain: dapp.chain,
+                        actions: dapp.id
+                    }
+                }),
+                thead: [
+                    t('common.appname_lbl'),
+                    t('common.origin_lbl'),
+                    t('common.account_lbl'),
+                    t('common.chain_lbl'),
+                    t('common.actions_lbl')
+                ],
+                tbody: ['appName', 'origin', 'displayString', 'chain', {slot: 'actions'}]
+            };
+        }
+    });
+
+    async function downloadBackup() {
+        let walletName = store.getters['WalletStore/getWalletName'];
+        let accounts = JSON.stringify(store.getters['AccountStore/getAccountList'].slice());
+
+        ipcRenderer.send(
+            "downloadBackup",
+            {
+                walletName: walletName,
+                accounts: accounts
+            }
+        );
+    }
+
+    async function deleteDapp(dapp_id) {
+        await store.dispatch('OriginStore/removeApp', dapp_id);
+        dapps.value = fetchDapps();
+    }
+</script>
+
 <template>
     <div class="bottom p-0">
         <div class="content">
             <div class="settings mt-3">
                 <p class="mb-1 font-weight-bold">
-                    {{ $t('settings_lbl') }}
+                    {{ t('common.settings_lbl') }}
                 </p>
             </div>
             <div class="dapp-list mt-2">
                 <p class="mb-2 font-weight-bold small">
-                    <u>{{ $t('dapps_lbl') }}</u>
+                    <u>{{ t('common.dapps_lbl') }}</u>
                 </p>
-                <table class="table small table-striped table-sm">
-                    <thead>
-                        <tr>
-                            <th
-                                class="align-middle"
+                <span v-if="tableData">
+                    <ui-table
+                        :data="tableData.data"
+                        :thead="tableData.thead"
+                        :tbody="tableData.tbody"
+                        style="height: 180px; overflow-y: scroll;"
+                    >
+                        <template #actions="{ data }">
+                            <ui-button
+                                raised
+                                @click="deleteDapp(data.actions)"
                             >
-                                {{ $t('appname_lbl') }}
-                            </th>
-                            <th
-                                class="align-middle"
-                            >
-                                {{ $t('origin_lbl') }}
-                            </th>
-                            <th
-                                class="align-middle"
-                            >
-                                {{ $t('account_lbl') }}
-                            </th>
-                            <th
-                                class="align-middle"
-                            >
-                                {{ $t('chain_lbl') }}
-                            </th>
-                            <th
-                                class="align-middle"
-                            >
-                                {{ $t('actions_lbl') }}
-                            </th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <tr v-if="dapps.length==0">
-                            <td
-                                colspan="6"
-                                class="align-center"
-                            >
-                                <em>{{ $t('no_dapps_linked') }}</em>
-                            </td>
-                        </tr>
-                        <tr
-                            v-for="dapp in dapps"
-                            :key="dapp.id"
-                        >
-                            <td
-                                class="align-middle"
-                            >
-                                {{ dapp.appName }}
-                            </td>
-                            <td
-                                class="align-middle"
-                            >
-                                {{ dapp.origin }}
-                            </td>
-                            <td
-                                class="align-middle"
-                                v-html="getDisplayString(dapp.account_id, dapp.chain)"
-                            >
-                            </td>
-                            <td
-                                class="align-middle"
-                            >
-                                {{ dapp.chain }}
-                            </td>
-                            <td
-                                class="align-middle"
-                            >
-                                <button
-                                    class="btn btn-sm btn-danger btn-block"
-                                    type="button"
-                                    @click="deleteDapp(dapp.id)"
-                                >
-                                    {{ $t('delete_btn') }}
-                                </button>
-                            </td>
-                        </tr>
-                    </tbody>
-                </table>
+                                {{ t('common.delete_btn') }}
+                            </ui-button>
+                        </template>
+                    </ui-table>
+                </span>
+                <span v-else>
+                    <em>{{ t('common.no_dapps_linked') }}</em>
+                </span>
             </div>
             <div class="backup mt-2 mb-4">
                 <p class="mb-2 font-weight-bold small">
-                    <u>{{ $t('backup_lbl') }}</u>
+                    <u>{{ t('common.backup_lbl') }}</u>
                 </p>
-                <div class="row px-4">
-                    <div class="col-12">
+                <ui-grid class="row px-4">
+                    <ui-grid-cell
+                        class="largeHeader"
+                        columns="12"
+                    >
                         <p class="small text-justify">
-                            {{ $t('backup_txt') }}
+                            {{ t('common.backup_txt') }}
                         </p>
-                    </div>
-                    <div class="col-3" />
-                    <div class="col-6">
-                        <button
-                            class="btn btn-sm btn-info btn-block"
+                    </ui-grid-cell>
+                    <ui-grid-cell columns="3" />
+                    <ui-grid-cell columns="6">
+                        <ui-button
+                            class="step_btn"
                             type="button"
+                            outlined
                             @click="downloadBackup"
                         >
-                            {{ $t('backup_btn') }}
-                        </button>
-                    </div>
-                    <div class="col-3" />
-                </div>
+                            {{ t('common.backup_btn') }}
+                        </ui-button>
+                    </ui-grid-cell>
+                    <ui-grid-cell columns="3" />
+                </ui-grid>
             </div>
         </div>
-        <Actionbar />
+        <router-link
+            :to="'/dashboard'"
+            replace
+        >
+            <ui-button
+                outlined
+                class="step_btn"
+            >
+                Exit settings menu
+            </ui-button>
+        </router-link>
     </div>
 </template>
-
-<script>
-    import { EventBus } from "../lib/event-bus.js";
-    import Actionbar from "./actionbar";
-    import {remote} from "electron";
-    import RendererLogger from "../lib/RendererLogger";
-    import path from 'path';
-    import { getBackup } from "../lib/SecureRemote";
-    import fs from 'fs';
-    import {formatAccount} from "../lib/formatter";
-
-    const logger = new RendererLogger();
-
-    export default {
-        name: "Settings",
-        i18nOptions: { namespaces: ["common"] },
-        components: {Actionbar},
-        data() {
-            return {
-
-            };
-        },
-        computed: {
-            dapps() {
-                let dapps=[];
-                for (let i=0;i< this.$store.state.AccountStore.accountlist.length;i++) {
-                    let apps=this.$store.getters['OriginStore/walletAccessibleDapps'](this.$store.state.AccountStore.accountlist[i].accountID,this.$store.state.AccountStore.accountlist[i].chain);
-                    if (typeof apps!='undefined') {
-                        dapps=dapps.concat(apps);
-                    }
-                }
-                return dapps;
-            }
-        },
-        watch: {},
-        created() {},
-        async mounted() {
-            logger.debug("Settings Mounted");
-            await this.$store.dispatch("OriginStore/loadApps");
-        },
-        methods: {
-            downloadBackup: async function () {
-
-                EventBus.$emit("popup", "load-start");
-                const dialog=remote.dialog;
-                const app=remote.app;
-                let remoteUrl="BeetBackup-"+this.$store.state.WalletStore.wallet.name+'-'+(new Date().toISOString().slice(0,10))+".beet";
-                let toLocalPath = path.resolve(app.getPath("desktop"), remoteUrl);
-                let userChosenPath = dialog.showSaveDialog({ defaultPath: toLocalPath });
-                if(userChosenPath) {
-                    let accounts=JSON.stringify({wallet: this.$store.state.WalletStore.wallet.name, accounts: this.$store.state.AccountStore.accountlist.slice()});
-                    let backup=await getBackup(accounts);
-                    if (backup) {
-                        fs.writeFileSync(userChosenPath,backup);
-                    }
-                    EventBus.$emit("popup", "load-end");
-                }else{
-                    EventBus.$emit("popup", "load-end");
-                }
-            },
-            deleteDapp: async function(dapp_id) {
-                await this.$store.dispatch('OriginStore/removeApp', dapp_id);
-            },
-            getDisplayString: function(accountID,chain) {
-                let account = this.$store.state.AccountStore.accountlist.find(x => { return (x.accountID==accountID && x.chain==chain);});
-                return formatAccount(account, true);
-            }
-        }
-    };
-</script>
