@@ -5,16 +5,10 @@ let _blockchains = {
         chainId: '4018d7844c78f6a6c41c6a552b898022310fc5dec06da467ee7905a8dad512c8',
         nodeList: [
             {
-                url: "wss://eu.nodes.bitshares.ws"
+                url: "wss://node.xbts.io/ws"
             },
-            {
-                url: "wss://newyork.bitshares.im/ws"
-            },
-            {
+             {
                 url: "wss://api.bts.mobi/ws"
-            },
-            {
-                url: "wss://api.iamredbar.com/ws"
             },
             {
                 url: "wss://nexus01.co.uk/ws"
@@ -23,13 +17,7 @@ let _blockchains = {
                 url: "wss://dex.iobanker.com/ws"
             },
             {
-                url: "wss://node.xbts.io/ws"
-            },
-            {
                 url: "wss://api.dex.trading/"
-            },
-            {
-                url: "wss://api-bts.liondani.com/ws"
             },
             {
                 url: "wss://api.bitshares.bhuz.info/ws"
@@ -72,6 +60,39 @@ let _blockchains = {
             }
         ]
     },
+    BTC: {
+        coreSymbol: 'BTC',
+        name: 'Bitcoin',
+        chainId: null,
+        nodeList: [
+            {
+                url: "https://blockchain.info/rawaddr/",
+                push: "https://blockchain.info/pushtx",
+                location: "Unknown",
+                region: "Unknown",
+                country: "Luxembourg",
+                operator: "Blockchain Luxembourg S.A.",
+                contact: "https://www.blockchain.com"
+            }
+        ]
+    },
+    BTC_TEST: {
+        coreSymbol: 'BTC',
+        name: 'Bitcoin',
+        chainId: null,
+        testnet: true,
+        nodeList: [
+            {
+                url: "https://testnet.blockchain.info/rawaddr/",
+                push: "https://testnet-api.smartbit.com.au/v1/blockchain/pushtx",
+                location: "Unknown",
+                region: "Unknown",
+                country: "Luxembourg",
+                operator: "Blockchain Luxembourg S.A.",
+                contact: "https://www.blockchain.com"
+            }
+        ]
+    }
     /*STEEM: {
         coreSymbol: 'STM',
         name: 'Steem',
@@ -131,40 +152,7 @@ let _blockchains = {
                 contact: "Greymass"
             }
         ]
-    },*/
-    BTC: {
-        coreSymbol: 'BTC',
-        name: 'Bitcoin',
-        chainId: null,
-        nodeList: [
-            {
-                url: "https://blockchain.info/rawaddr/",
-                push: "https://blockchain.info/pushtx",
-                location: "Unknown",
-                region: "Unknown",
-                country: "Luxembourg",
-                operator: "Blockchain Luxembourg S.A.",
-                contact: "https://www.blockchain.com"
-            }
-        ]
     },
-    BTC_TEST: {
-        coreSymbol: 'BTC',
-        name: 'Bitcoin',
-        chainId: null,
-        testnet: true,
-        nodeList: [
-            {
-                url: "https://testnet.blockchain.info/rawaddr/",
-                push: "https://testnet-api.smartbit.com.au/v1/blockchain/pushtx",
-                location: "Unknown",
-                region: "Unknown",
-                country: "Luxembourg",
-                operator: "Blockchain Luxembourg S.A.",
-                contact: "https://www.blockchain.com"
-            }
-        ]
-    }/*,
     BNB_TEST: {
         coreSymbol: 'BNB',
         name: 'BinanceChain',
