@@ -1,9 +1,9 @@
-import getBlockchain from "../lib/blockchains/blockchainFactory";
+import getBlockchainAPI from "../lib/blockchains/blockchainFactory";
 
 export default class operations {
 
     static async generate(data, account_id) {
-        const blockchain = getBlockchain();
+        const blockchain = getBlockchainAPI();
         let operation = await blockchain.getOperation(data, account_id);
         return {
             type: operation.type,
