@@ -1,9 +1,6 @@
-import {
-    ipcRenderer,
-} from 'electron';
-import { ec as EC } from "elliptic";
-import CryptoJS from 'crypto-js';
-const ec = new EC('secp256k1');
+import { ipcRenderer } from 'electron';
+import * as secp from "@noble/secp256k1";
+import sha256 from "crypto-js/sha256.js";
 
 class proover {
     constructor() {
@@ -11,39 +8,98 @@ class proover {
     }
 
     async regen() {
-        this.key = ec.genKeyPair();
-        let publicKey = this.key.getPublic().encode('hex');
-        ipcRenderer.send('key', publicKey);
+        this.key = secp.utils.randomPrivateKey();
+        let pubk;
+        try {
+          pubk = await secp.getPublicKey(this.key);
+        } catch (error) {
+          console.error(error);
+          return;
+        }
+        this.pubk = pubk;
     }
 
-    sign(data) {
-        return this.key.sign(CryptoJS.SHA256(data).toString()).toDER();
+    async sign(data) {
+        let msgHash;
+        try {
+          msgHash = await sha256(data).toString();
+        } catch (error) {
+          console.log(error);
+          return;
+        }
+
+        let signedMessage;
+        try {
+          signedMessage = await secp.sign(
+            msgHash,
+            this.key,
+            {der: true, extraEntropy: true}
+          )
+        } catch (error) {
+          console.log(error);
+          return;
+        }
+
+        return {
+          signedMessage: signedMessage,
+          msgHash: msgHash,
+          pubk: this.pubk
+        };
     }
 }
 
 const proof = new proover();
 
-export const getKey = (enc_key) => {
-    return new Promise(resolve => {
-        ipcRenderer.removeAllListeners('decrypt');
-        ipcRenderer.once('decrypt', (event, arg) => {            
+export const getKey = async (enc_key) => {
+    return new Promise(async (resolve, reject) => {
+        ipcRenderer.removeAllListeners('decrypt_success');
+        ipcRenderer.removeAllListeners('decrypt_fail');
+
+        ipcRenderer.once('decrypt_success', (event, arg) => {
+            console.log('decrypt_success')
             resolve(arg);
         });
-        ipcRenderer.send('decrypt', {
-            data: enc_key,
-            sig: proof.sign('decrypt')
-        })
+
+        ipcRenderer.once('decrypt_fail', (event, arg) => {
+            console.log('decrypt_fail')
+            reject('decrypt_fail');
+        });
+
+        let signature = await getSignature('decrypt');
+        if (!signature) {
+          console.log('Signature failure')
+          reject('signature failure');
+        }
+
+        let isValid;
+        try {
+          isValid = await secp.verify(
+            signature.signedMessage,
+            signature.msgHash,
+            signature.pubk
+          );
+        } catch (error) {
+          console.log(error);
+        }
+
+        if (isValid) {
+          console.log("Was valid, proceeding to decrypt");
+          ipcRenderer.send('decrypt', {data: enc_key});
+        } else {
+          console.log('invalid signature')
+          reject('invalid signature');
+        }
     })
 }
-export const getBackup = (data) => {
-    return new Promise(resolve => {
-        ipcRenderer.removeAllListeners('backup');
-        ipcRenderer.once('backup', (event, arg) => {            
-            resolve(arg);
-        });
-        ipcRenderer.send('backup', {
-            data: data,
-            sig: proof.sign('backup')
-        })
-    })
+
+export const getSignature = async (data) => {
+  let signature;
+  try {
+    signature = await proof.sign(data);
+  } catch (error) {
+    console.log(error);
+    return;
+  }
+
+  return signature;
 }
