@@ -1,18 +1,18 @@
+import mitt from 'mitt';
+const emitter = mitt();
 import store from '../store/index.js';
-import { EventBus } from "./event-bus.js";
 
 export default class TimeoutService {
-
     constructor(timeout) {
-        this.timeout=timeout;
-        EventBus.$on('keyaccess', () => { 
+        this.timeout = timeout;
+        emitter.on('keyaccess', () => {
             this.resetTimeout();
         });
-        this.timer=setTimeout(this.expireKeys,this.timeout);
+        this.timer = setTimeout(this.expireKeys, this.timeout);
     }
     resetTimeout() {
         clearTimeout(this.timer);
-        this.timer=setTimeout(this.expireKeys,this.timeout);
+        this.timer = setTimeout(this.expireKeys, this.timeout);
     }
     expireKeys() {
         store.dispatch('AccountStore/expireKeys');
