@@ -4,58 +4,361 @@ import {
     PrivateKey,
     PublicKey,
     TransactionBuilder,
-    Signature,
-    FetchChain
+    Signature
 } from "bitsharesjs";
+import * as Socket from "simple-websocket";
+
 import RendererLogger from "../RendererLogger";
 import {formatAsset, humanReadableFloat} from "../assetUtils";
 const logger = new RendererLogger();
 
+
+const permission_flags = {
+    charge_market_fee: 0x01 /**< an issuer-specified percentage of all market trades in this asset is paid to the issuer */,
+    white_list: 0x02 /**< accounts must be whitelisted in order to hold this asset */,
+    override_authority: 0x04 /**< issuer may transfer asset back to himself */,
+    transfer_restricted: 0x08 /**< require the issuer to be one party to every transfer */,
+    disable_force_settle: 0x10 /**< disable force settling */,
+    global_settle: 0x20 /**< allow the bitasset issuer to force a global settling -- this may be set in permissions, but not flags */,
+    disable_confidential: 0x40 /**< allow the asset to be used with confidential transactions */,
+    witness_fed_asset: 0x80 /**< allow the asset to be fed by witnesses */,
+    committee_fed_asset: 0x100 /**< allow the asset to be fed by the committee */,
+    lock_max_supply: 0x200, ///< the max supply of the asset can not be updated
+    disable_new_supply: 0x400, ///< unable to create new supply for the asset
+    disable_mcr_update: 0x800, ///< the bitasset owner can not update MCR, permission only
+    disable_icr_update: 0x1000, ///< the bitasset owner can not update ICR, permission only
+    disable_mssr_update: 0x2000, ///< the bitasset owner can not update MSSR, permission only
+    disable_bsrm_update: 0x4000, ///< the bitasset owner can not update BSRM, permission only
+    disable_collateral_bidding: 0x8000 ///< Can not bid collateral after a global settlement
+};
+
+const uia_permission_mask = [
+    "charge_market_fee",
+    "white_list",
+    "override_authority",
+    "transfer_restricted",
+    "disable_confidential"
+];
+
+/**
+ * 
+ * @param {String} mask 
+ * @param {Boolean} isBitAsset 
+ * @returns Object
+ */
+function getFlagBooleans(mask, isBitAsset = false) {
+    let booleans = {
+        charge_market_fee: false,
+        white_list: false,
+        override_authority: false,
+        transfer_restricted: false,
+        disable_force_settle: false,
+        global_settle: false,
+        disable_confidential: false,
+        witness_fed_asset: false,
+        committee_fed_asset: false,
+        lock_max_supply: false,
+        disable_new_supply: false,
+        disable_mcr_update: false,
+        disable_icr_update: false,
+        disable_mssr_update: false,
+        disable_bsrm_update: false,
+        disable_collateral_bidding: false
+    };
+    
+    if (mask === "all") {
+        for (let flag in booleans) {
+            if (
+                !isBitAsset &&
+                uia_permission_mask.indexOf(flag) === -1
+            ) {
+                delete booleans[flag];
+            } else {
+                booleans[flag] = true;
+            }
+        }
+        return booleans;
+    }
+    
+    for (let flag in booleans) {
+        if (
+            !isBitAsset &&
+            uia_permission_mask.indexOf(flag) === -1
+        ) {
+            delete booleans[flag];
+        } else {
+            if (mask & permission_flags[flag]) {
+                booleans[flag] = true;
+            }
+        }
+    }
+    
+    return booleans;
+}
+
+
 export default class BitShares extends BlockchainAPI {
 
-    _connect(nodeToConnect = null) {
-        return new Promise((resolve, reject) => {
-            if (nodeToConnect == null) {
-                nodeToConnect = this.getNodes()[0].url;
-            }
-            if (this._isConnected) {
-                Apis.close().then(() => {
-                    this._isConnected = false;
-                    Apis.instance(
-                        nodeToConnect,
-                        true,
-                        10000,
-                        {enableCrypto: false, enableOrders: false},
-                        // no use in firing reject because it might happen at any time in the future after connecting!
-                        this._connectionFailed.bind(this, null, nodeToConnect, "Connection closed")
-                    ).init_promise.then(() => {
-                        this._connectionEstablished(resolve, nodeToConnect);
-                    }).catch(this._connectionFailed.bind(this, reject, nodeToConnect));
+    /*
+     * Signing a Bitshares NFT with the user's account.
+     * @param {string} key
+     * @param {String} nft_object
+     * @returns {Promise}
+     */
+    signNFT(key, nft_object) {
+        return new Promise((resolve,reject) => {
+            let updatedObject = JSON.parse(nft_object);
+            updatedObject.sig_pubkey_or_address = this.getPublicKey(key);
+            try {
+                resolve({
+                    key: this.getPublicKey(key),
+                    signed: updatedObject,
+                    signature: this._signString(key, JSON.stringify(updatedObject))
                 });
-            } else {
-                Apis.instance(
-                    nodeToConnect,
-                    true,
-                    10000,
-                    {enableCrypto: false, enableOrders: false},
-                    // no use in firing reject because it might happen at any time in the future after connecting!
-                    this._connectionFailed.bind(this, null, nodeToConnect, "Connection closed")
-                ).init_promise.then(() => {
-                    this._connectionEstablished(resolve, nodeToConnect);
-                }).catch(this._connectionFailed.bind(this, reject, nodeToConnect));
+            } catch (error) {
+                console.log(error)
+                reject(error);
             }
         });
     }
 
-    _needsReconnecting() {
-        if (this._isConnected) {
-            let _isConnectedToTestnet = Apis.instance().url.indexOf("testnet") !== -1;
-            return _isConnectedToTestnet !== this._isTestnet();
-        } else {
-            return false;
-        }
+    /**
+     * Test a wss url for successful connection.
+     * @param {String} url
+     * @returns {Object}
+     */
+    _testConnection(url) {
+      return new Promise(async (resolve, reject) => {
+        let before = new Date();
+        let beforeTS = before.getTime();
+
+        let socket = new Socket(url);
+        socket.on('connect', () => {
+          let now = new Date();
+          let nowTS = now.getTime();
+          socket.destroy();
+          return resolve({ url: url, lag: nowTS - beforeTS });
+        });
+
+        socket.on('error', (error) => {
+          console.log(error);
+          socket.destroy();
+          return resolve(null);
+        });
+      });
     }
 
+  /**
+   * Test the wss nodes, return latencies and fastest url.
+   * @returns {Promise}
+   */
+    async _testNodes() {
+      return new Promise(async (resolve, reject) => {
+
+          let urls = this.getNodes().map(node => node.url);
+
+          let filteredURLS = urls.filter(url => {
+            if (!this._tempBanned || !this._tempBanned.includes(url)) {
+              return true;
+            }
+          });
+
+          return Promise.all(filteredURLS.map(url => this._testConnection(url)))
+          .then((validNodes) => {
+            let filteredNodes = validNodes.filter(x => x);
+            if (filteredNodes.length) {
+              let sortedNodes = filteredNodes.sort((a, b) => a.lag - b.lag);
+              console.log(`Fastest node: ${sortedNodes[0].url}`);
+
+              let now = new Date();
+              return resolve({
+                node: sortedNodes[0].url,
+                latencies: sortedNodes,
+                timestamp: now.getTime()
+              });
+            } else {
+              console.error("No valid BTS WSS connections established; Please check your internet connection.")
+              return reject();
+            }
+          })
+          .catch(error => {
+            console.log(error);
+          })
+
+
+      });
+
+    }
+
+    /**
+     * Fetch a working node to connect to, using bitsharesws-js manager class
+     * Unused code - manager class failed to disconnect fast enough.
+     * @returns {Promise}
+     */
+     /*
+    _fetchValidNode() {
+        return new Promise((resolve, reject) => {
+            let urls = this.getNodes().map(node => node.url);
+
+            let filteredURLS = urls.filter(url => {
+              if (!this._tempBanned || !this._tempBanned.includes(url)) {
+                return true;
+              }
+            });
+
+            let connectionManager = new Manager({
+                url: filteredURLS[0],
+                urls: filteredURLS,
+                closeCb: res => {
+                  console.log(res);
+                },
+                optionalApis: {enableOrders: true},
+                urlChangeCallback: url => {
+                    console.log("urlChangeCallback:", url);
+                }
+            })
+
+            connectionManager
+            .checkConnections()
+            .then(res => {
+                let urls = Object.keys(res);
+                let ascLagNodes = urls.map(url => {
+                                    return { url: url, lag: res[url] };
+                                  }).sort((a, b) => a.lag - b.lag);
+                //console.log("best node: ", ascLagNodes[0]);
+                let now = new Date();
+                resolve({
+                  node: ascLagNodes[0].url,
+                  latencies: ascLagNodes,
+                  timestamp: now.getTime()
+                })
+            })
+            .catch(err => {
+                console.log("doLatencyUpdate error", err);
+                reject();
+            })
+        });
+    }
+    */
+
+    /*
+     * Check if the connection needs reestablished (placeholder replacement)
+     * @returns {Boolean}
+     */
+    async _needsNewConnection() {
+        return new Promise(async (resolve, reject) => {
+            if (
+                !this._isConnected ||
+                !this._isConnectedToNode ||
+                !this._nodeLatencies
+            ) {
+                return resolve(true);
+            }
+      
+            if (this._isTestnet()) {
+                let _isConnectedToTestnet = Apis.instance().url.indexOf("testnet") !== -1;
+                return resolve(_isConnectedToTestnet !== this._isTestnet());
+            }
+    
+            let testConnection = await this._testConnection(this._isConnectedToNode);
+            let connectionResult = testConnection && testConnection.url ? false : true;
+            return resolve(connectionResult);
+        });
+    }
+
+    /*
+     * Establish a connection
+     * @param {String} nodeToConnect
+     * @param {Promise} resolve
+     * @param {Promise} reject
+     * @returns {String}
+     */
+    _establishConnection(nodeToConnect, resolve, reject) {
+        if (!nodeToConnect) {
+          this._connectionFailed(reject, '', 'No node url')
+        }
+
+        Apis.instance(
+            nodeToConnect,
+            true,
+            4000,
+            {enableCrypto: false, enableOrders: false},
+            console.log('Initial WSS Connection closed')
+        ).init_promise
+        .then((res) => {
+          console.log("established connection:", res[0].network);
+          this._connectionEstablished(resolve, nodeToConnect);
+        })
+        .catch(error => {
+          console.log(error);
+          this._connectionFailed(reject, nodeToConnect, error)
+        });
+    }
+
+    /*
+     * Connect to the Bitshares blockchain. (placeholder replacement)
+     * @param {String||null} nodeToConnect
+     * @returns {String}
+     */
+    _connect(nodeToConnect = null) {
+        return new Promise((resolve, reject) => {
+
+            if (nodeToConnect) {
+                //console.log(`nodetoconnect: ${nodeToConnect}`)
+                return this._establishConnection(nodeToConnect, resolve, reject);
+            }
+
+            if (this._isConnected && this._isConnectedToNode && !nodeToConnect) {
+                //console.log(`isConnected: ${this._isConnectedToNode}`)
+                return this._connectionEstablished(resolve, this._isConnectedToNode);
+            }
+
+            let diff;
+            if (this._nodeCheckTime) {
+                let now = new Date();
+                let nowTS = now.getTime();
+                diff = Math.abs(Math.round((nowTS - this._nodeCheckTime) / 1000));
+            }
+
+            if (!nodeToConnect && (!this._nodeLatencies || diff && diff > 360)) {
+                // initializing the blockchain
+                console.log('Checking node connections')
+                return this._testNodes().then((res) => {
+                  this._node = res.node;
+                  this._nodeLatencies = res.latencies;
+                  this._nodeCheckTime = res.timestamp;
+                  console.log(`Establishing connection to ${res.node}`);
+                  return this._establishConnection(res.node, resolve, reject);
+                })
+                .catch(error => {
+                  console.log(error);
+                  return this._connectionFailed(reject, '', 'Node test fail');
+                })
+            } else if (!nodeToConnect && this._nodeLatencies) {
+              // blockchain has previously been initialized
+              let filteredNodes = this._nodeLatencies
+                                  .filter(item => {
+                                    if (!this._tempBanned.includes(item.url)) {
+                                      return true;
+                                    }
+                                  });
+
+              this._nodeLatencies = filteredNodes;
+              if (!filteredNodes || !filteredNodes.length) {
+                return this._connectionFailed(reject, '', 'No working nodes');
+              }
+
+              this._node = filteredNodes[0].url;
+              return this._establishConnection(filteredNodes[0].url, resolve, reject);
+            }
+
+        });
+    }
+
+    /*
+     * Returns an array of default import options. (placeholder replacement)
+     * @returns {Array}
+     */
     getImportOptions() {
         return [
             {
@@ -77,25 +380,48 @@ export default class BitShares extends BlockchainAPI {
         ];
     }
 
-    getAccount(accountName) {
-        return new Promise((resolve, reject) => {
-            this.ensureConnection().then(() => {
-                Apis.instance().db_api()
+    /*
+     * Retrieve the contents of the provided account name from the blockchain API
+     * @param {String} accountName
+     * @returns {Object} parsedAccount
+     */
+    async getAccount(accountName) {
+        return new Promise(async (resolve, reject) => {
+              this.ensureConnection().then(() => {
+                Apis.instance()
+                    .db_api()
                     .exec("get_full_accounts", [[accountName], false])
-                    .then(res => {
-                        res[0][1].account.active.public_keys = res[0][1].account.active.key_auths;
-                        res[0][1].account.owner.public_keys = res[0][1].account.owner.key_auths;
-                        res[0][1].account.memo = {public_key: res[0][1].account.options.memo_key};
-                        res[0][1].account.balances = res[0][1].balances;
-                        resolve(res[0][1].account);
+                    .then(response => {
+                        if (!response || !response.length || !response[0].length) {
+                            console.log('Failed to query BTS blockchain');
+                            return reject('Failed to query BTS blockchain');
+                        }
+
+                        let parsedAccount = response[0][1].account;
+                        parsedAccount.active.public_keys = parsedAccount.active.key_auths;
+                        parsedAccount.owner.public_keys = parsedAccount.owner.key_auths;
+                        parsedAccount.memo = {public_key: parsedAccount.options.memo_key};
+                        parsedAccount.balances = response[0][1].balances;
+                        return resolve(parsedAccount);
                     })
-                    .catch((err) => {
-                        reject(err);
-                    });
-            }).catch(reject);
+                    .catch(error => {
+                        console.log(`get_full_accounts: ${error}`);
+                        return this._connectionFailed(reject, this._node, error)
+                    })
+              })
+              .catch(error => {
+                  console.log(`ensureConnection: ${error}`);
+                  reject(error);
+              })
+
         });
     }
 
+    /*
+     * Get the associated Bitshares account name from the provided account ID
+     * @param {String} accountId
+     * @returns {String}
+     */
     _getAccountName(accountId) {
         return new Promise((resolve, reject) => {
             this.ensureConnection().then(() => {
@@ -108,6 +434,11 @@ export default class BitShares extends BlockchainAPI {
         });
     }
 
+    /*
+     * Retrieve an asset object from a provided asset symbol or ID
+     * @param {String} assetSymbolOrId
+     * @returns {Object}
+     */
     _resolveAsset(assetSymbolOrId) {
         return new Promise((resolve, reject) => {
             this.ensureConnection().then(() => {
@@ -120,48 +451,49 @@ export default class BitShares extends BlockchainAPI {
         });
     }
 
+    /*
+     * Retrieve an asset object from a provided asset symbol or ID
+     * @param {String} assetSymbolOrId
+     * @returns {Object}
+     */
     getAsset(assetSymbolOrId) {
-        if (this._isTestnet()) {
-            if (assetSymbolOrId == "1.3.0") {
-                return {
-                    asset_id: "1.3.0",
-                    symbol: "TEST",
-                    precision: 5
-                };
-            } else {
-                return null;
-            }
-        } else {
-            if (assetSymbolOrId == "1.3.0") {
-                return {
-                    asset_id: "1.3.0",
-                    symbol: "BTS",
-                    precision: 5
-                };
-            } else if (assetSymbolOrId == "1.3.121") {
-                return {
-                    asset_id: "1.3.121",
-                    symbol: "bitUSD",
-                    precision: 4
-                };
-            } else if (assetSymbolOrId == "1.3.113") {
-                return {
-                    asset_id: "1.3.113",
-                    symbol: "bitCNY",
-                    precision: 4
-                };
-            } else if (assetSymbolOrId == "1.3.120") {
-                return {
-                    asset_id: "1.3.120",
-                    symbol: "bitEUR",
-                    precision: 4
-                };
-            } else {
-                return null;
-            }
-        }
+        return new Promise((resolve, reject) => {
+
+          if (this._isTestnet()) {
+              if (assetSymbolOrId == "1.3.0") {
+                  return resolve({
+                      asset_id: "1.3.0",
+                      symbol: "TEST",
+                      precision: 5
+                  });
+              } else {
+                  // TODO: Provide testnet bitshares lookup
+                  return reject(null);
+              }
+          }
+
+          this.ensureConnection().then(() => {
+              Apis.instance().db_api().exec("lookup_asset_symbols", [[assetSymbolOrId]]).then((asset_objects) => {
+                  if (!asset_objects.length || !asset_objects[0]) {
+                    return resolve(null);
+                  }
+
+                  let retrievedAsset = asset_objects[0];
+                  return resolve({
+                      asset_id: retrievedAsset.id,
+                      symbol: retrievedAsset.symbol,
+                      precision: retrievedAsset.precision
+                  });
+              }).catch(reject);
+          }).catch(reject);
+        });
     }
 
+    /*
+     * Retrieve the Bitshares balances for the provided account name.
+     * @param {String} accountName
+     * @returns {Array} balances
+     */
     getBalances(accountName) {
         return new Promise((resolve, reject) => {
             // getAccount has already ensureConnection
@@ -195,16 +527,30 @@ export default class BitShares extends BlockchainAPI {
                     }
                     resolve(balances);
                 });
-            }).catch(reject);
+            }).catch((error) => {
+                console.log(error);
+                reject(error);
+            });
         });
     }
 
+    /*
+     * Retrieve the pubkey from a provided BTS private key
+     * @param {String} privateKey
+     * @returns {String} publicKey
+     */
     getPublicKey(privateKey) {
         return PrivateKey.fromWif(privateKey)
             .toPublicKey()
             .toString(this._getCoreSymbol());
     }
 
+    /*
+     * Map operational data for incoming Bitshares vote actions
+     * Retrieves committee members, witnesses and worker proposal objects.
+     * @param {Object} incoming
+     * @returns {Object}
+     */
     mapOperationData(incoming) {
         return new Promise((resolve, reject) => {
             this.ensureConnection().then(() => {
@@ -233,7 +579,10 @@ export default class BitShares extends BlockchainAPI {
                                             incoming.params.id,
                                         vote_id: objdata[0].vote_id
                                     });
-                                }).catch(err => reject(err));
+                                }).catch(error => {
+                                  console.log(error);
+                                  reject(error)
+                                });
                                 break;
                             case "6":
                                 Apis.instance().db_api().exec(
@@ -248,7 +597,10 @@ export default class BitShares extends BlockchainAPI {
                                             incoming.params.id,
                                         vote_id: objdata[0].vote_id
                                     });
-                                }).catch(err => reject(err));
+                                }).catch(error => {
+                                  console.log(error);
+                                  reject(error)
+                                });
                                 break;
                             case "14":
                                 Apis.instance().db_api().exec(
@@ -268,15 +620,26 @@ export default class BitShares extends BlockchainAPI {
                                             objextradata[0].name,
                                         vote_id: objdata[0].vote_for
                                     });
-                                }).catch(err => reject(err));
+                                }).catch(error => {
+                                  console.log(error);
+                                  reject(error)
+                                });
                                 break;
                         }
-                    }).catch(err => reject(err));
+                    }).catch(error => {
+                      console.log(error);
+                      reject(error)
+                    });
                 }
             }).catch(reject);
         });
     }
 
+    /*
+     * Parse incoming and return a readied transaction builder instance
+     * @param {Class||Object} incoming
+     * @returns {Class} TransactionBuilder
+     */
     _parseTransactionBuilder(incoming) {
         if (incoming instanceof TransactionBuilder) {
             return incoming;
@@ -308,6 +671,12 @@ export default class BitShares extends BlockchainAPI {
         throw "Reconstruction of TransactionBuilder failed";
     }
 
+    /*
+     * Parse incoming and return a readied transaction builder instance
+     * @param {Class||Object} operation
+     * @param {String} key
+     * @returns {Class} TransactionBuilder
+     */
     sign(operation, key) {
         return new Promise((resolve, reject) => {
             this.ensureConnection().then(() => {
@@ -317,85 +686,102 @@ export default class BitShares extends BlockchainAPI {
                     tr.update_head_block()
                 ]).then(() => {
                     let privateKey = PrivateKey.fromWif(key);
-                    tr.add_signer(privateKey, privateKey.toPublicKey().toPublicKeyString(this._getCoreSymbol()));
+                    tr.add_signer(
+                      privateKey,
+                      privateKey.toPublicKey().toPublicKeyString(this._getCoreSymbol()));
                     tr.finalize().then(() => {
                         tr.sign();
                         resolve(tr);
                     }).catch(reject);
                 });
-            }).catch(err => reject(err));
+            }).catch(error => {
+              console.log(error);
+              reject(error)
+            });
         });
     }
 
+    /*
+     * Broadcast a transaction on the Bitshares blockchain.
+     * @param {Class||Object} transaction
+     * @returns {Object} broadcastResult
+     */
     broadcast(transaction) {
         return new Promise((resolve, reject) => {
             this.ensureConnection().then(() => {
                 transaction = this._parseTransactionBuilder(transaction);
                 transaction.broadcast().then(id => {
                     resolve(id);
-                }).catch(err => reject(err));
+                }).catch(error => {
+                  console.log(error);
+                  reject(error)
+                });
             }).catch(reject);
         });
     }
 
+    /*
+     * Retrieve operation from input data and account combination
+     * @param {Object} data
+     * @param {Object} account
+     * @returns {Object}
+     */
     getOperation(data, account) {
-        let account_id = account.id;
         return new Promise((resolve, reject) => {
             this.ensureConnection().then(() => {
-                let operation = {
-                    type: null,
-                    data: null
-                };
-                let api = Apis.instance();
-                switch (data.action) {
-                    case 'vote': {
-                        operation.type = 'account_update';
-                        api.db_api().exec(
-                            "get_objects",
-                            [[account_id]]
-                        ).then((accounts) => {
-                            let updateObject = {
-                                account: account_id
-                            };
 
-                            let account = accounts[0];
+                if (data.action === 'vote') {
+                  let accountID;
+                  try {
+                    accountID = account.accountID;
+                  } catch (error) {
+                    console.log(error)
+                  }
 
-                            let new_options = account.options;
+                  Apis.instance().db_api().exec("get_objects", [[accountID]]).then((accounts) => {
 
-                            if (new_options.votes.findIndex(item => item == data.vote_id) !== -1) {
-                                resolve({
-                                   vote_id: data.vote_id,
-                                   notingToDo: true
-                                });
-                            }
+                      let new_options = accounts[0].options;
+                      if (new_options.votes.findIndex(item => item == data.vote_id) !== -1) {
+                          resolve({
+                             vote_id: data.vote_id,
+                             nothingToDo: true
+                          });
+                      }
 
-                            new_options.votes.push(data.vote_id);
-                            new_options.votes = new_options.votes.sort((a, b) => {
-                                let a_split = a.split(":");
-                                let b_split = b.split(":");
-                                return (
-                                    parseInt(a_split[1], 10) - parseInt(b_split[1], 10)
-                                );
-                            });
-                            updateObject.new_options = new_options;
-                            operation.data = updateObject;
-                            resolve(operation);
-                        }).catch(err => {
-                            reject(err);
-                        });
+                      new_options.votes.push(data.vote_id);
+                      new_options.votes = new_options.votes.sort((a, b) => {
+                          let a_split = a.split(":");
+                          let b_split = b.split(":");
+                          return (
+                              parseInt(a_split[1], 10) - parseInt(b_split[1], 10)
+                          );
+                      });
+                      resolve({
+                        data: {
+                            account: accountID,
+                            new_options: new_options
+                        },
+                        type: 'account_update'
+                      });
 
-                    }
-                    break;
-                    default: {
-                        operation.type = 'transfer';
-                        operation.data = data;
-                        resolve(operation);
-                    }
+                  }).catch(error => {
+                      console.log(error)
+                      reject(error);
+                  });
+
+                } else {
+                  resolve({data: data, type: 'transfer'});
                 }
             }).catch(reject);
         });
     }
 
+    /*
+     * Signs a string using a provided private key.
+     * @param {String} key
+     * @param {String} string
+     * @returns {String} hexString
+     */
     _signString(key, string) {
         let signature = Signature.signBuffer(
             string,
@@ -404,6 +790,13 @@ export default class BitShares extends BlockchainAPI {
         return signature.toHex();
     }
 
+    /*
+     * Verifies that the signature is the BTS public key signed string.
+     * @param {String} signature (hex)
+     * @param {String} publicKey
+     * @param {String} string
+     * @returns {Boolean}
+     */
     _verifyString(signature, publicKey, string) {
         let _PublicKey = PublicKey;
         let sig = Signature.fromHex(signature);
@@ -414,29 +807,62 @@ export default class BitShares extends BlockchainAPI {
         );
     }
 
+    /*
+     * Broadcast a transfer operation on the Bitshares blockchain.
+     * @param {String} key
+     * @param {String} from
+     * @param {String} to
+     * @param {String} amount
+     * @param {String} memo
+     * @returns {Object} transfer result
+     */
     async transfer(key, from, to, amount, memo = null) {
         if (!amount.amount || !amount.asset_id) {
             throw "Amount must be a dict with amount and asset_id as keys"
         }
         from = await this.getAccount(from);
         to = await this.getAccount(to);
-        let operation = {
-            type: "transfer",
-            data: {
-                fee: {
-                    amount: 0,
-                    asset_id: "1.3.0"
-                },
-                from: from.id,
-                to: to.id,
-                amount: amount,
-                memo: memo == null ? undefined : memo
-            }
-        };
-        let transaction = await this.sign(operation, key);
-        return await this.broadcast(transaction);
+
+        let transaction;
+        try {
+          transaction = await this.sign(
+            {
+                type: "transfer",
+                data: {
+                    fee: {
+                        amount: 0,
+                        asset_id: "1.3.0"
+                    },
+                    from: from.id,
+                    to: to.id,
+                    amount: amount,
+                    memo: memo == null ? undefined : memo
+                }
+            },
+            key
+          );
+        } catch (error) {
+          console.log(error);
+          throw "Could not sign operation with Bitshares key";
+        }
+
+        let broadcastResult;
+        try {
+          broadcastResult = await this.broadcast(transaction);
+        } catch (error) {
+          console.log(error);
+          throw "Could not broadcast signed Bitshares transaction";
+        }
+
+        return broadcastResult
     }
 
+    /*
+     * Return an appropriate Bitshares blockchain explorer link.
+     * Warning: Opens dangerously, be cautious adding alt explorers.
+     * @param {Object} object
+     * @returns {String}
+     */
     getExplorer(object) {
         if (object.accountName) {
             return "https://blocksights.info/#/accounts/" + object.accountID;
@@ -451,6 +877,10 @@ export default class BitShares extends BlockchainAPI {
         }
     }
 
+    /*
+     * Returns the required keys for sign up input forms
+     * @returns {Object}
+     */
     getSignUpInput() {
         return {
             active: false,
@@ -459,27 +889,78 @@ export default class BitShares extends BlockchainAPI {
         }
     }
 
+    /*
+     * Returns the remaining nodes sorted by asc latency
+     * @returns {Array}
+     */
+    getLatencies() {
+      return this._nodeLatencies ?? [];
+    }
+
+    /*
+     * Returns a visualization for the input data.
+     * TODO: Requires refactor
+     * @param {String||Class||Object} thing
+     * @returns {String}
+     */
     async visualize(thing) {
         if (typeof thing == "string" && thing.startsWith("1.2.")) {
             // resolve id to name
-            return this._getAccountName(thing);
+            return await this._getAccountName(thing);
         }
+
         let operations = [];
         let tr = this._parseTransactionBuilder(thing);
-        console.log("Visualizing " + tr);
+
+        console.log(tr.operations)
+
         for (let i = 0; i < tr.operations.length; i++) {
             let operation = tr.operations[i];
             if (operation[0] == 0) {
-                let from = await this._getAccountName(operation[1].from);
-                let to = await this._getAccountName(operation[1].to);
-                let asset = await this._resolveAsset(operation[1].amount.asset_id);
+                let from;
+                try {
+                  from = await this._getAccountName(operation[1].from);
+                } catch (error) {
+                  console.log(error);
+                  return;
+                }
+
+                let to;
+                try {
+                  to = await this._getAccountName(operation[1].to);
+                } catch (error) {
+                  console.log(error);
+                  return;
+                }
+
+                let asset;
+                try {
+                  asset = await this._resolveAsset(operation[1].amount.asset_id);
+                } catch (error) {
+                  console.log(error);
+                  return;
+                }
 
                 operations.push(
                     from + " &#9657; " + formatAsset(operation[1].amount.amount, asset.symbol, asset.precision) + " &#9657; " + to
                 )
             } else if (operation[0] == 25) {
-                let to = await this._getAccountName(operation[1].authorized_account);
-                let asset = await this._resolveAsset(operation[1].withdrawal_limit.asset_id);
+                let to;
+                try {
+                  to = await this._getAccountName(operation[1].authorized_account);
+                } catch (error) {
+                  console.log(error);
+                  return;
+                }
+
+                let asset;
+                try {
+                  asset = await this._resolveAsset(operation[1].withdrawal_limit.asset_id);
+                } catch (error) {
+                  console.log(error);
+                  return;
+                }
+
                 let period = operation[1].withdrawal_period_sec / 60 / 60 / 24;
                 operations.push(
                     "Direct Debit Authorization\n" +
@@ -487,16 +968,51 @@ export default class BitShares extends BlockchainAPI {
                     " Take " + formatAsset(operation[1].withdrawal_limit.amount, asset.symbol, asset.precision) + " every " + period + " days, for " + operation[1].periods_until_expiration + " periods"
                 )
             } else if (operation[0] == 33) {
-                let owner = await this._getAccountName(operation[1].owner);
-                let asset = await this._resolveAsset(operation[1].amount.asset_id);
+                let owner;
+                try {
+                  owner = await this._getAccountName(operation[1].owner);
+                } catch (error) {
+                  console.log(error);
+                  return;
+                }
+
+                let asset;
+                try {
+                  asset = await this._resolveAsset(operation[1].amount.asset_id);
+                } catch (error) {
+                  console.log(error);
+                  return;
+                }
+
                 operations.push(
                     "Vesting Balance\n" +
                     " Claim " + formatAsset(operation[1].amount.amount, asset.symbol, asset.precision) + " from balance " + operation[1].vesting_balance
                 )
             } else if (operation[0] == 1) {
-                let seller = await this._getAccountName(operation[1].seller);
-                let buy = await this._resolveAsset(operation[1].min_to_receive.asset_id);
-                let sell = await this._resolveAsset(operation[1].amount_to_sell.asset_id);
+                let seller;
+                try {
+                  seller = await this._getAccountName(operation[1].seller);
+                } catch (error) {
+                  console.log(error);
+                  return;
+                }
+
+                let buy;
+                try {
+                  buy = await this._resolveAsset(operation[1].min_to_receive.asset_id);
+                } catch (error) {
+                  console.log(error);
+                  return;
+                }
+
+                let sell;
+                try {
+                  sell = await this._resolveAsset(operation[1].amount_to_sell.asset_id);
+                } catch (error) {
+                  console.log(error);
+                  return;
+                }
+
                 let fillOrKill = operation[1].amount_to_sell.fill_or_kill;
 
                 let price = humanReadableFloat(operation[1].amount_to_sell.amount, sell.precision)
@@ -508,17 +1024,120 @@ export default class BitShares extends BlockchainAPI {
                     " Buy: " + formatAsset(operation[1].min_to_receive.amount, buy.symbol, buy.precision) + "\n" +
                     " Price: " + price.toPrecision(6) + " " + sell.symbol + "/" +  buy.symbol
                 )
+            } else if (operation[0] == 10 || operation[0] == 11) {
+                // Create or Update an asset
+
+
+                let op = operation[1];
+
+                let asset;
+                if (operation[0] == 11) {
+                    // fetch asset to update
+                    try {
+                      asset = await this._resolveAsset(operation[1].asset_to_update);
+                    } catch (error) {
+                      console.log(error);
+                      return;
+                    }
+                }
+
+                let symbol = asset ? asset.symbol : op.symbol;
+                let precision = asset ? asset.precision : op.precision;
+                let is_prediction_market = asset ? asset.is_prediction_market : op.is_prediction_market;
+
+                let options = operation[0] == 10 ? op.common_options : op.new_options;
+
+                let max_supply = options.max_supply;
+                let market_fee_percent = options.market_fee_percent;
+                let max_market_fee = options.max_market_fee;
+                let isBitasset = op.bitasset_opts ? true : false;
+
+                let issuer_permissions = getFlagBooleans(options.issuer_permissions, isBitasset);
+                let flags = getFlagBooleans(options.flags, isBitasset);
+
+                let cer_base_amount = options.core_exchange_rate.base.amount;
+                let cer_base_asset_id = options.core_exchange_rate.base.asset_id;
+                let cer_quote_amount = options.core_exchange_rate.quote.amount;
+                let cer_quote_asset_id = options.core_exchange_rate.quote.asset_id; 
+
+                let whitelist_authorities = options.whitelist_authorities;
+                let blacklist_authorities = options.blacklist_authorities;
+                let whitelist_markets = options.whitelist_markets;
+                let blacklist_markets = options.blacklist_markets;
+
+                let description = JSON.parse(options.description);
+                let nft_object = description ? description.nft_object : null;
+
+                let initialString = operation[0] == 10 ? "Issuing an asset \n" : "Updating an asset \n"
+                let operationString =  initialString +
+                                        `Symbol: ${symbol}\n` +
+                                        `main: ${description.main}\n` +
+                                        `market: ${description.market}\n` +
+                                        `short_name: ${description.short_name}\n` +
+                                        `Precision: ${precision}\n` +
+                                        `max supply: ${max_supply}\n` +
+                                        `market_fee_percent: ${market_fee_percent}\n` +
+                                        `max_market_fee: ${max_market_fee}\n\n` +
+                                        //
+                                        `Core exchange rates:\n` +
+                                        `Base amount: ${cer_base_amount}\n` +
+                                        `Base asset id: ${cer_base_asset_id}\n` +
+                                        `Quote amount: ${cer_quote_amount}\n` +
+                                        `Quote asset id: ${cer_quote_asset_id}\n` +
+                                        `whitelist_authorities: ${whitelist_authorities}\n` +
+                                        `blacklist_authorities: ${blacklist_authorities}\n` +
+                                        `whitelist_markets: ${whitelist_markets}\n` +
+                                        `blacklist_markets: ${blacklist_markets}\n` +
+                                        `is_prediction_market: ${is_prediction_market}\n\n` +
+                                        //
+                                        `Permissions:\n` + 
+                                        `charge_market_fee: ${issuer_permissions["charge_market_fee"]}\n` +
+                                        `white_list: ${issuer_permissions["white_list"]}\n` +
+                                        `override_authority: ${issuer_permissions["override_authority"]}\n` +
+                                        `transfer_restricted: ${issuer_permissions["transfer_restricted"]}\n` +
+                                        `disable_confidential: ${issuer_permissions["disable_confidential"]}\n\n` +
+                                        //
+                                        `Flags:\n` +
+                                        `charge_market_fee: ${flags["charge_market_fee"]}\n` +
+                                        `white_list: ${flags["white_list"]}\n` +
+                                        `override_authority: ${flags["override_authority"]}\n` +
+                                        `transfer_restricted: ${flags["transfer_restricted"]}\n` +
+                                        `disable_confidential: ${flags["disable_confidential"]}\n\n`;
+                
+                if (isBitasset) {
+                    operationString += `Bitasset info: \n`;
+                    operationString += `feed_lifetime_sec: ${op.bitasset_opts.feed_lifetime_sec}\n`;
+                    operationString += `force_settlement_delay_sec: ${op.bitasset_opts.force_settlement_delay_sec}\n`;
+                    operationString += `force_settlement_offset_percent: ${op.bitasset_opts.force_settlement_offset_percent}\n`;
+                    operationString += `maximum_force_settlement_volume: ${op.bitasset_opts.maximum_force_settlement_volume}\n`;
+                    operationString += `minimum_feeds: ${op.bitasset_opts.minimum_feeds}\n`;
+                    operationString += `short_backing_asset: ${op.bitasset_opts.short_backing_asset}\n`;
+                }
+
+                if (nft_object) {
+                    operationString += `NFT Contents: \n`;
+                    operationString += `acknowledgements: ${nft_object.acknowledgements}\n`;
+                    operationString += `artist: ${nft_object.artist}\n`;
+                    operationString += `attestation: ${nft_object.attestation}\n`;
+                    operationString += `holder_license: ${nft_object.holder_license}\n`;
+                    operationString += `license: ${nft_object.license}\n`;
+                    operationString += `narrative: ${nft_object.narrative}\n`;
+                    operationString += `title: ${nft_object.title}\n`;
+                    operationString += `tags: ${nft_object.tags}\n`;
+                    operationString += `type: ${nft_object.type}\n`;
+                }
+
+                operations.push(operationString);
             }
         }
-        if (operations.length == 0) {
+
+        if (!operations.length) {
             return false;
         }
+
         let header = operations.length == 1 ? "" : "Transaction\n";
 
-
-        return `<pre class="text-left custom-content">
-<code>${header}${operations.join('\n')}
-</code></pre>`;
+        return `${header}${operations.join('\n')}`;
     }
 
 }
