@@ -1,77 +1,123 @@
 import { blockchains } from "../../config/config.js";
 
-let apiCache = {}
-
 import BitShares from "./BitShares"
 import TUSC from "./TUSC"
-import Steem from "./Steem"
 import store from "../../store";
+import Bitcoin from "./Bitcoin";
+
+/*
+import Steem from "./Steem"
 import WhaleShares from "./WhaleShares";
 import EOSmainnet from "./EOSmainnet";
 import TLOS from "./TLOS";
-import Bitcoin from "./Bitcoin";
 import Binance from "./Binance";
+*/
 
-export default function getBlockchainAPI(chain = null) {
+let bts,bts_test,tusc,steem,wls,eos,tlos,btc,btc_test,bnb,bnb_test;
+
+export default function getBlockchainAPI(chain = null, node = null) {
     if (chain == null) {
-        // ask store
-        chain = store.state.AccountStore.accountlist[store.state.AccountStore.selectedIndex].chain
+        chain = store.getters['AccountStore/getChain'];
     }
+
+    let config;
+    try {
+      config = blockchains[chain];
+    } catch (error) {
+      console.log(error);
+      return;
+    }
+
     if (chain == "BTS") {
-        if (!apiCache.BTS) {
-            apiCache.BTS = new BitShares(blockchains[chain]);
+        if (!bts) {
+            try {
+              bts = new BitShares(config, node);
+            } catch (error) {
+              console.log(error);
+              return;
+            }
         }
-        return apiCache.BTS;
+        return bts;
     } else if (chain == "BTS_TEST") {
-        if (!apiCache.BTS_TEST) {
-            apiCache.BTS_TEST = new BitShares(blockchains[chain]);
+        if (!bts_test) {
+            try {
+              bts_test = new BitShares(config, node);
+            } catch (error) {
+              console.log(error);
+              return;
+            }
         }
-        return apiCache.BTS_TEST;
+        return bts_test;
     } else if (chain == "TUSC") {
-        if (!apiCache.TUSC) {
-            apiCache.TUSC = new TUSC(blockchains[chain]);
+        if (!tusc) {
+            try {
+              tusc = new TUSC(config, node);
+            } catch (error) {
+              console.log(error);
+              return;
+            }
         }
-        return apiCache.TUSC;
-    } else if (chain == "STEEM" || chain == "STM") {
-        if (!apiCache.STEEM) {
-            apiCache.STEEM = new Steem(blockchains["STEEM"]);
+        return tusc;
+    } else if (chain == "BTC") {
+        if (!btc) {
+            try {
+              btc = new Bitcoin(config, node);
+            } catch (error) {
+              console.log(error);
+              return;
+            }
         }
-        return apiCache.STEEM;
+        return btc;
+    } else if (chain == "BTC_TEST") {
+        if (!btc_test) {
+            try {
+              btc_test = new Bitcoin(config, node);
+            } catch (error) {
+              console.log(error);
+              return;
+            }
+        }
+        return btc_test;
+    }/* else if (chain == "STEEM" || chain == "STM") {
+        if (!steem) {
+            try {
+              steem = new Steem(config, node);
+            } catch (error) {
+              console.log(error);
+              return;
+            }
+        }
+        return steem;
     } else if (chain == "WLS") {
-        if (!apiCache.WLS) {
-            apiCache.WLS = new WhaleShares(blockchains[chain]);
+        if (!wls) {
+            try {
+              wls = new WhaleShares(config, node);
+            } catch (error) {
+              console.log(error);
+              return;
+            }
         }
-        return apiCache.WLS;
+        return wls;
     } else if (chain == "EOS") {
         if (!apiCache.EOS) {
-            apiCache.EOS = new EOSmainnet(blockchains[chain]);
+            apiCache.EOS = new EOSmainnet(config, node);
         }
         return apiCache.EOS;
     } else if (chain == "TLOS") {
         if (!apiCache.TLOS) {
-            apiCache.TLOS = new TLOS(blockchains[chain]);
+            apiCache.TLOS = new TLOS(config, node);
         }
         return apiCache.TLOS;
-    } else if (chain == "BTC") {
-        if (!apiCache.BTC) {
-            apiCache.BTC = new Bitcoin(blockchains[chain]);
-        }
-        return apiCache.BTC;
-    } else if (chain == "BTC_TEST") {
-        if (!apiCache.BTC_TEST) {
-            apiCache.BTC_TEST = new Bitcoin(blockchains[chain]);
-        }
-        return apiCache.BTC_TEST;
     } else if (chain == "BNB") {
         if (!apiCache.BNB) {
-            apiCache.BNB = new Binance(blockchains[chain]);
+            apiCache.BNB = new Binance(config, node);
         }
         return apiCache.BNB;
     } else if (chain == "BNB_TEST") {
         if (!apiCache.BNB_TEST) {
-            apiCache.BNB_TEST = new Binance(blockchains[chain]);
+            apiCache.BNB_TEST = new Binance(config, node);
         }
         return apiCache.BNB_TEST;
-    }
+    }*/
 
 }
