@@ -1,4 +1,5 @@
-import VueRouter from 'vue-router'
+import { createRouter, createWebHashHistory } from 'vue-router'
+
 import HeaderGuest from "../components/header-guest";
 import HeaderSmall from "../components/header-small";
 import Start from "../components/start";
@@ -6,8 +7,9 @@ import Dashboard from "../components/dashboard";
 import Restore from "../components/restore";
 import AddAccount from "../components/add-account";
 import Settings from "../components/settings";
+import Popups from "../components/popups";
 
-const router = new VueRouter({
+const router = createRouter({
   routes: [{
       path: '/',
       components: {
@@ -49,7 +51,15 @@ const router = new VueRouter({
         default: Dashboard,
         header: HeaderSmall
       }
+    },
+    {
+      path: '/modal',
+      components: {
+        default: Popups,
+        header: HeaderSmall
+      }
     }
-  ]
+  ],
+  history: createWebHashHistory()
 });
 export default router;
